removing database dependencies
