+++ go.mod
@@ -3,13 +3,10 @@ module m3u-stream-merger
 go 1.23.0
 
 require (
+	github.com/goccy/go-json v0.10.3
 	github.com/gosimple/slug v1.14.0
-	github.com/redis/go-redis/v9 v9.7.0
+	github.com/klauspost/compress v1.17.11
 	github.com/robfig/cron/v3 v3.0.1
 )
 
-require (
-	github.com/cespare/xxhash/v2 v2.3.0 // indirect
-	github.com/dgryski/go-rendezvous v0.0.0-20200823014737-9f7001d12a5f // indirect
-	github.com/gosimple/unidecode v1.0.1 // indirect
-)
+require github.com/gosimple/unidecode v1.0.1 // indirect
