+++ store/filters.go
@@ -1,7 +1,6 @@
-package m3u
+package store
 
 import (
-	"m3u-stream-merger/database"
 	"m3u-stream-merger/utils"
 	"regexp"
 )
@@ -10,7 +9,7 @@ var includeFilters [][]string
 var excludeFilters [][]string
 var filtersInitialized bool
 
-func checkFilter(stream database.StreamInfo) bool {
+func checkFilter(stream StreamInfo) bool {
 	if !filtersInitialized {
 		excludeFilters = [][]string{
 			utils.GetFilters("EXCLUDE_GROUPS"),
