+++ tests/proxy_test.go
@@ -1,13 +1,11 @@
-package proxy
+package tests
 
 import (
 	"bytes"
-	"context"
 	"io"
 	"log"
-	"m3u-stream-merger/database"
-	"m3u-stream-merger/m3u"
-	"m3u-stream-merger/updater"
+	"m3u-stream-merger/handlers"
+	"m3u-stream-merger/store"
 	"net/http"
 	"net/http/httptest"
 	"os"
@@ -16,35 +14,17 @@ import (
 )
 
 func TestStreamHandler(t *testing.T) {
-	os.Setenv("REDIS_ADDR", "127.0.0.1:6379")
-	os.Setenv("REDIS_PASS", "")
-	os.Setenv("REDIS_DB", "0")
-
-	db, err := database.InitializeDb()
-	if err != nil {
-		t.Errorf("InitializeDb returned error: %v", err)
-	}
-
-	err = db.ClearDb()
-	if err != nil {
-		t.Errorf("ClearDb returned error: %v", err)
-	}
-
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
 	os.Setenv("M3U_URL_1", "https://gist.githubusercontent.com/sonroyaalmerol/de1c90e8681af040924da5d15c7f530d/raw/06844df09e69ea278060252ca5aa8d767eb4543d/test-m3u.m3u")
 	os.Setenv("INCLUDE_GROUPS_1", "movies")
 
-	_, err = updater.Initialize(ctx)
+	err := store.DownloadM3USource(1)
 	if err != nil {
-		t.Errorf("Updater returned error: %v", err)
+		t.Errorf("Downloader returned error: %v", err)
 	}
 
-	streamChan := db.GetStreams()
-	streams := []database.StreamInfo{}
+	streams := store.GetStreams()
 
-	for stream := range streamChan {
+	for _, stream := range streams {
 		streams = append(streams, stream)
 	}
 
@@ -52,7 +32,7 @@ func TestStreamHandler(t *testing.T) {
 	m3uW := httptest.NewRecorder()
 
 	func() {
-		m3u.Handler(m3uW, m3uReq)
+		handlers.M3UHandler(m3uW, m3uReq)
 	}()
 
 	m3uResp := m3uW.Result()
@@ -60,15 +40,17 @@ func TestStreamHandler(t *testing.T) {
 		t.Errorf("Playlist Route - Expected status code %d, got %d", http.StatusOK, m3uResp.StatusCode)
 	}
 
+	cm := store.NewConcurrencyManager()
+
 	for _, stream := range streams {
 		log.Printf("Stream (%s): %v", stream.Title, stream)
-		genStreamUrl := strings.TrimSpace(m3u.GenerateStreamURL("", stream))
+		genStreamUrl := strings.TrimSpace(store.GenerateStreamURL("", stream))
 
 		req := httptest.NewRequest("GET", genStreamUrl, nil)
 		w := httptest.NewRecorder()
 
 		// Call the handler function
-		Handler(w, req)
+		handlers.StreamHandler(w, req, cm)
 
 		// Check the response status code
 		resp := w.Result()
